q9.cpp

+++ level1/q9.cpp
@@ -6,7 +6,7 @@ int main()
 {
 
   
-    int Age;
+    short int Age;
         cout << "please enter your age: \n";
         cin >> Age;
         cout << "after 5 years you will be " << Age + 5<<" years old\n";
